Handle single letter + apostrophe better
In the special case of a single letter plus apostrophe leading a word,
often we're either looking at a non-English word with abbreviated
article (usually starting with a consonant) or a name (usually starting
with a vowel). We make an educated guess and assume the former should
be lower case and the latter upper case.

Handles issue #5.