Defer lowering all caps until recursion is done
Consider `titlecase("RED-BLACK NEWSPAPER TITLE")`:

Because `RED-BLACK` is hyphenated a null callback will have three chances:

  1. `RED-BLACK`
  2. `red`
  3. `black`

It is odd that the argument to 2 and 3 is lower cased when
the original input was not. (A similar thing happens for
words with /'s.) This commit defers lower-casing until the
hyphenated phrase is fully resolved.

Addresses Issue #9.