Update supported / tested versions -py26 +py36 ~pypy
Py26 has been deprecated for a while and throws an error uploading
data to coveralls, that's a good enough reason for me to drop it.

Travis's default PyPy is also too old, but we can just bump to something
newer and keep that around.

Py36 has been released, so let's add that to the mix as well.