setup.py: Read version from __init__.py to avoid importing the module
The module import happens before setup() is executed, and therefore,
none of the modules specified in 'setup_requires' (or 'install_requires',
for that matter) are loaded at that moment, causing the import to fail
with ModuleNotFoundError.

Following the first recommendation from the 'Single-sourcing the
package version' section of the Python Package User Guide [1], in this
commit we switch to reading the version string from the '__init__.py'
file instead of attempting to import the module.

[1]: https://packaging.python.org/guides/single-sourcing-package-version/

---

Also includes updates in patch at
https://github.com/ppannuto/python-titlecase/pull/54#issuecomment-643048143