Always capitalize 'Mc'-prefixed small words in compound word groups
Although there are no real surnames that start with ‘Mc’ and continues
with a word matched by the SMALL_WORDS regex, it feels weird for
a string like ‘mcto/mcby’ to be title-cased into ‘McTo/Mcby’, and not
‘McTo/McBy’, even though both ‘to’ and ‘by’ are small words—just
because the small_first_last argument is set to False in the slash-
or hyphen-induced recursion. Hence this fix.